Fixed massive bug, added test, ran against web-p2 more
Summary:
Only the first in the list of rules to evaluate was actually being evaluated. This was due to a massive function where my return was indented one level in. python....

Added a new test and a sink for it

Ran it a bunch and cleaned up a few things

Differential Revision: https://code.uberinternal.com/D498391